refactor: 코드리뷰내용 반영
