+++ backend/src/main/java/com/example/rabbithell/domain/auth/dto/request/FullJwtRequest.java
@@ -1,6 +1,7 @@
 package com.example.rabbithell.domain.auth.dto.request;
 
 public record FullJwtRequest(
-	String cloverName
+	String cloverName,
+	Long kingdomId
 ) {
 }
