+++ backend/src/main/java/com/example/rabbithell/infrastructure/security/jwt/JwtProperties.java
@@ -4,32 +4,28 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
 import lombok.Getter;
-import lombok.Setter;
 
 @Component
 @ConfigurationProperties(prefix = "spring.jwt")
 @Getter
-@Setter
 public class JwtProperties {
+
     private String secret;
-    private Token token;
+    private final Token token = new Token();
 
     @Getter
-    @Setter
     public static class Token {
         private String header;
         private String prefix;
-        private Access access;
-        private Refresh refresh;
+        private final Access access = new Access();
+        private final Refresh refresh = new Refresh();
 
         @Getter
-        @Setter
         public static class Access {
             private long minute;
         }
 
         @Getter
-        @Setter
         public static class Refresh {
             private long minute;
         }
